Kill off remaining git invocations and custom SQL.
Now the code works purely using the database and does it in a
SQL-variant agnostic manner. This will run on a normal webstack
where the webserver host can be torn down at any time.

Previously, the git repository was local to the webserver and
thus article bodies would not persist server rebuilds.

This set up should also be faster and much more optimizable.